Fix errors with regexes using `|`
We are applying the `\G` prefix to the regex which means it must match
at the provided index, but if the regex has an OR in it, that prefix only
applies to the first option of the OR. Thus a subsequent match could
occur anywhere in the string.

Fix is to wrap the whole regex in a non-capturing group, and then apply
the `\G` prefix to the front of that.